requirements.txt
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blobfile==3.0.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.2
contourpy==1.3.0
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
distro==1.9.0
drawsvg @ file:///home/conda/feedstock_root/build_artifacts/drawsvg_1735836149330/work
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.104.1
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.58.0
fqdn==1.5.1
fsspec==2024.6.1
ftfy==6.3.1
h11==0.16.0
hf-xet==1.1.8
-e git+https://github.com/aminshabani/house_diffusion.git@24ff52cd945ea2b425a4652e826c9d88d7c0330d#egg=house_diffusion
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.13.0
joblib==1.5.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
loguru==0.7.2
lxml==5.4.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==3.1.3
mpi4py @ file:///croot/mpi4py_1671223370575/work
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.4.4
notebook_shim==0.2.4
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
nvitop==1.5.1
openai==2.30.0
opencv-python==4.11.0.86
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
pygraphviz @ file:///home/conda/feedstock_root/build_artifacts/pygraphviz_1700370314207/work
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
python-multipart==0.0.6
pytorch-fid==0.3.0
pytorch-tabnet==4.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
Send2Trash==1.8.3
sh==2.2.2
Shapely==1.8.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.27.0
svgwrite==1.4.3
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.21.4
tomli==2.2.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
triton==3.2.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
zipp==3.21.0
